+++ NEWS.md
@@ -22,6 +22,8 @@
 
 * Added the `"add_overall.tbl_svysummary-arg:col_label"` theme element to control the default value for `add_overall.tbl_svysummary(col_label)` argument. (#2354)
 
+* Added theme elements `add_ci.tbl_summary-arg:method`, and `add_ci.tbl_svysummary-arg:method` to control the default value for `add_ci.tbl_summary(method)` / `add_ci.tbl_svysummary(method)` argument. (#2356)
+
 # gtsummary 2.4.0
 
 ### New Features and Functions